Add files via upload
This notebook is all about analysing stochiastic signals , in order to obtain some information hidden in them. This data is a sample data from fuel cell device. More information about fuel cell can be found here. (https://www.hydrogenics.com/technology-resources/hydrogen-technology/fuel-cells/), which works on hydrogen. Data science is slowly making a progress in this area in order to predict the parameters and analyse signals.

Why data science is helpful ?

This is helpful because a fuel cell is highly complex device, it is not important to understand all parameters and to get a link between them. So we develop a black box model and then try to find relations using other algorithms such as neural network/ logistics/ SVM etc.

Here we will be exploring the signals. In order to process the signal,lets explore the data first. Signal processing is important to understand the reality/hidden information of the signals in specific domains. Not every signal are subjected to remove noise. In this tutorial, I will explore the signal in time domain and frequency domain and then to analyse the signals.